fixes include/exclude issues with special characters
While excludes are resolved by rsync and can simply be quoted, includes
will be expanded by a shell and therefor need special treatment.
As we don't know if a given item is a proper path, possibly containing
spaces, or a pattern we have to do our own escaping.
For this we escape ALL characters not matching [a-zA-Z0-9,._+:@%/\-*]
Hence include pattern will only support wildcard expansion.